feat: engine optimization, advanced modes, and web playtest system
- Optimized SolverEngine with 3.7x speedup and MDAS precedence.\n- Added Tunnel Vision and Permutations modes with pre-verification.\n- Implemented Dart Shelf web playtest backend with AOT support.\n- Created CalibrationScreen for background math offloading.\n- Refined Jeopardy distribution logic and added Restart Match feature.\n- Configured .gitignore and .geminiignore for playtest artifacts.

+++ packages/game_engine/lib/src/career_manager.dart
@@ -79,7 +79,7 @@ class CareerManager {
     _unlockedAchievements.add(id);
   }
 
-  void recordRoundPerformance({required double secondsToSubmit, required int proximityToTarget}) {
+  void recordRoundPerformance({required double secondsToSubmit, required double proximityToTarget}) {
     _avgSpeedSeconds = ((_avgSpeedSeconds * _roundsTracked) + secondsToSubmit) / (_roundsTracked + 1);
     _avgAccuracy = ((_avgAccuracy * _roundsTracked) + proximityToTarget) / (_roundsTracked + 1);
     _roundsTracked++;
@@ -110,7 +110,7 @@ class CareerManager {
     }
   }
 
-  void recordMultiplayerPerformance({required double secondsToSubmit, required int proximityToTarget, required String opponentName, int eloShift = 0}) {
+  void recordMultiplayerPerformance({required double secondsToSubmit, required double proximityToTarget, required String opponentName, int eloShift = 0}) {
     _avgSpeedSeconds = ((_avgSpeedSeconds * _roundsTracked) + secondsToSubmit) / (_roundsTracked + 1);
     _avgAccuracy = ((_avgAccuracy * _roundsTracked) + proximityToTarget) / (_roundsTracked + 1);
     _roundsTracked++;
